% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\docType{data}
\name{aggregation}
\alias{aggregation}
\title{Aggregation}
\format{
A data frame with 788 rows and 2 columns, \code{x1} and \code{x2}.
}
\usage{
aggregation
}
\description{
A synthetic data set of two-dimensional points created by Gionis et al. to
demonstrate clustering aggregation.
}
\references{
A. Gionis, H. Mannila, and P. Tsaparas, Clustering aggregation.
ACM Transactions on Knowledge Discovery from Data (TKDD), 2007. 1(1): p. 1-30.
}
\keyword{datasets}
