% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pald_functions.R
\name{any_isolated}
\alias{any_isolated}
\title{Any isolated}
\usage{
any_isolated(c)
}
\arguments{
\item{c}{A matrix of cohesion values (see \code{\link{cohesion_matrix}}).}
}
\value{
Logical, indicating whether any points are isolated.
}
\description{
Checks for isolated points.
}
