% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognate_dist.R
\docType{data}
\name{cognate_dist}
\alias{cognate_dist}
\title{Cognate Data Distance Matrix}
\format{
A \code{\link{dist}} object for 87 Indo-European languages.
}
\usage{
cognate_dist
}
\description{
A \code{\link{dist}} object describing distances between 87 Indo-European languages from
the perspective of cognates.
}
\details{
Cognate relationships from a collection of essential words were collected
from Dyen et al. and encoded in a 87x2665 binary matrix from which this
distance matrix was derived (using Euclidean distance).
}
\references{
I. Dyen, J. B. Kruskal, P. Black, An Indoeuropean
classification: A lexicostatistical experiment. Trans. Am. Phil. Soc. 82,
iii-132 (1992).
}
\keyword{datasets}
