% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pald_functions.R
\name{community_clusters}
\alias{community_clusters}
\title{Community clusters}
\usage{
community_clusters(c)
}
\arguments{
\item{c}{A \code{cohesion_matrix} object, a matrix of cohesion values
(see \code{\link{cohesion_matrix}}).}
}
\value{
A data frame with two columns:
\itemize{
\item \code{point}: The points from cohesion matrix \code{c}
\item \code{cluster}: The (community) cluster labels
}
}
\description{
Community clusters
}
\examples{
D <- dist(exdata2)
C <- cohesion_matrix(D)
community_clusters(C)
}
