% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata1.R
\docType{data}
\name{exdata1}
\alias{exdata1}
\title{Example Data 1}
\format{
A data frame with 8 rows and 2 variables.
}
\usage{
exdata1
}
\description{
A data set consisting of 8 points (in 2-dimensional Euclidean space)
to provide a simple illustrative example.  This data is displayed in Figure 1 in BMM22.
}
\keyword{datasets}
