% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata3.R
\docType{data}
\name{exdata3}
\alias{exdata3}
\title{Example Data 3}
\format{
A data frame with 240 rows and 2 variables.
}
\usage{
exdata3
}
\description{
A data set consisting of 240 points (in 2-dimensional Euclidean space)
to provide an illustrative example.  Points were generated from bivariate normal distributions with varying mean and variance (with covariance matrix cI). This data is displayed in Figure 4D in BMM22.
}
\keyword{datasets}
