% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_circles.R
\docType{data}
\name{noisy_circles}
\alias{noisy_circles}
\title{Noisy circles}
\format{
A dataframe with 500 rows and 2 columns, \code{x1} and \code{x2}.
}
\source{
\url{https://scikit-learn.org/stable/modules/clustering.html#clustering}
}
\usage{
noisy_circles
}
\description{
Noisy circles data generated from scikit-learn
}
\keyword{datasets}
