% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_moons.R
\docType{data}
\name{noisy_moons}
\alias{noisy_moons}
\title{Noisy moons}
\format{
A dataframe with 500 rows and 2 columns, \code{x1} and \code{x2}.
}
\source{
\url{https://scikit-learn.org/stable/modules/clustering.html#clustering}
}
\usage{
noisy_moons
}
\description{
Noisy moons data generated from scikit-learn
}
\keyword{datasets}
