% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pald_colors.R
\docType{data}
\name{pald_colors}
\alias{pald_colors}
\title{PaLD Color Palette}
\format{
A vector of 24 colors
}
\usage{
pald_colors
}
\description{
A vector of colors to use if comparing other clustering methods. These are
the default colors used in the plotting functions.
}
\keyword{datasets}
