% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pald_functions.R
\name{plot_community_graphs}
\alias{plot_community_graphs}
\title{Plot Community Graphs}
\usage{
plot_community_graphs(
  C,
  layout = NULL,
  show.labels = TRUE,
  only_strong = FALSE,
  emph_strong = 2,
  edge.width.factor = 50,
  vertex.lab = rownames(C),
  vertex.label.cex = 1,
  colors = NULL,
  vertex.color.vec = NULL,
  vertex.size = 1
)
}
\arguments{
\item{C}{A matrix of cohesion values (see \code{cohesion_matrix}).}

\item{layout}{A layout for the graph.  If none is specified, FR-graph drawing algorithm is used.}

\item{show.labels}{Set to FALSE to omit vertex labels (to display a subset of labels, use optional input "vertex.lab" and modify the label list).}

\item{only_strong}{Set to TRUE if only strong ties, G_strong, should be displayed; the default FALSE will show both strong (colored by connected component) and weak ties (in gray).}

\item{emph_strong}{The numeric factor by which the edge widths of strong ties are emphasized in the display; the default is 2.}

\item{edge.width.factor}{Modify to change displayed edge widths.}

\item{vertex.lab}{A vector containing label names.}

\item{vertex.label.cex}{A numeric value for modifying the vertex label size.}

\item{colors}{A vector of display colors, if none is given a default is provided.}

\item{vertex.color.vec}{A vector of colors names for coloring vertices.}

\item{vertex.size}{A numeric value for vertex size.}
}
\description{
Provides a plot of the community graphs, with connected components of the graph of strong ties colored by connected component.
}
\details{
Plots the community graph, G, with the sub-graph of strong ties emphasized and colored by connected component.  If no layout is provided, the Fructerman-Reingold (FR) graph drawing algorithm is used.
Note that the FR graph drawing algorithm provides a somewhat different layout each time it is run.  You can also access and save a given graph layout using \code{community_graphs(C)$layout}.
The example below shows how to display only a subset of vertex labels.

Note that the parameter \code{emph_strong} is for visualization purposes only and does not influence the network layout.
}
\examples{
C <- cohesion_matrix(dist(exdata1))
plot_community_graphs(C, emph_strong = 1, layout = exdata1)
plot_community_graphs(C, only_strong = TRUE)

C2<-cohesion_matrix(cognate_dist)
subset_lang_names<-rownames(C2)
subset_lang_names[sample(1:87, 60)]<-''
plot_community_graphs(C2, vertex.lab = subset_lang_names, vertex.size = 3)
}
