% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_cohesion_matrix.R
\docType{data}
\name{tissue_cohesion_matrix}
\alias{tissue_cohesion_matrix}
\title{Tissue Cohesion Matrix}
\format{
A cohesion matrix with 189 rows and 189 columns
}
\usage{
tissue_cohesion_matrix
}
\description{
A cohesion matrix created from a subset of tissue gene expression data from
the following papers:
\itemize{
\item http://www.ncbi.nlm.nih.gov/pubmed/17906632
\item http://www.ncbi.nlm.nih.gov/pubmed/21177656
\item http://www.ncbi.nlm.nih.gov/pubmed/24271388
obtained from the \strong{tissuesGeneExpression} bioconductor package.
}
}
\details{
The original data frame had 189 rows,  each with a corresponding tissue,
such as \code{colon}, \code{kidney} or \code{cerebellum}.
There were 22,215 columns corresponding to gene expression data from each of
these rows. This was then converted into a cohesion matrix.
}
\references{
M. Love and R. Irizarry. tissueGeneExpression. Bioconductor
Package
}
\keyword{datasets}
