% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata1.R
\docType{data}
\name{exdata1}
\alias{exdata1}
\title{Example data 1}
\format{
A data frame with 8 rows and 2 variables.
}
\usage{
exdata1
}
\description{
A dataset consisting of 8 points (in 2-dimensional Euclidean space)
for the purposes of example
}
\keyword{datasets}
