% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata1.R
\docType{data}
\name{exdata1}
\alias{exdata1}
\title{Example Data 1}
\format{
A data frame with 8 rows and 2 columns, \code{x1} and \code{x2}
}
\usage{
exdata1
}
\description{
A data set consisting of 8 points (in 2-dimensional Euclidean space)
to provide a simple illustrative example.  This data is displayed in Figure
1 in Berenhaut, Moore, and Melvin (2022).
}
\references{
K. S. Berenhaut, K. E. Moore, R. L. Melvin, A social perspective
on perceived distances reveals deep community structure. Proc. Natl.
Acad. Sci., 119(4), 2022.
}
\keyword{datasets}
