% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata2.R
\docType{data}
\name{exdata2}
\alias{exdata2}
\title{Example Data 2}
\format{
A data frame with 16 rows and 2 variables.
}
\usage{
exdata2
}
\description{
A data set consisting of 16 points (in 2-dimensional Euclidean space)
to provide an illustrative example.  This data is displayed in Figure 2 in BMM22.
}
\keyword{datasets}
