% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata2.R
\docType{data}
\name{exdata2}
\alias{exdata2}
\title{Example Data 2}
\format{
A data frame with 16 rows and 2 columns, \code{x1} and \code{x2}
}
\usage{
exdata2
}
\description{
A data set consisting of 16 points (in 2-dimensional Euclidean space)
to provide an illustrative example.  This data is displayed in Figure 2
in Berenhaut, Moore, and Melvin (2022).
}
\references{
K. S. Berenhaut, K. E. Moore, R. L. Melvin, A social perspective
on perceived distances reveals deep community structure. Proc. Natl.
Acad. Sci., 119(4), 2022.
}
\keyword{datasets}
