% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata3.R
\docType{data}
\name{exdata3}
\alias{exdata3}
\title{Example Data 3}
\format{
A data frame with 240 rows and 2 columns, \code{x1} and \code{x2}
}
\usage{
exdata3
}
\description{
A data set consisting of 240 points (in 2-dimensional Euclidean space)
to provide an illustrative example.  Points were generated from bivariate
normal distributions with varying mean and variance (with covariance
matrix cI). This data is displayed in Figure 4D in Berenhaut, Moore, and
Melvin (2022).
}
\references{
K. S. Berenhaut, K. E. Moore, R. L. Melvin, A social perspective
on perceived distances reveals deep community structure. Proc. Natl.
Acad. Sci., 119(4), 2022.
}
\keyword{datasets}
